implemented seperate dqn module and MHA state encoder
