+++ dqn/FINDER_encoder.py
@@ -1,0 +1,15 @@
+import tensorflow as tf
+
+class AGNNEncoder(tf.keras.layers.Layer):
+    def __init__(self, placeholder_dict, num_layers, node_embed_dim, weight_stddev, rate=0.1):
+        super(AGNNEncoder, self).__init__()
+        pass
+    def call(self, x):
+        pass
+
+class AGNNEncoderLayer(tf.keras.layers.Layer):
+    def __init__(self, placeholder_dict, num_layers, node_embed_dim, weight_stddev, rate=0.1):
+        super(AGNNEncoderLayer, self).__init__()
+        pass
+    def call(self, x):
+        pass
